프론트 테스트 위한 일부 설정 및 Board 기능 구현

+++ src/main/java/com/ssafeople/backend/domain/board/presentation/BoardController.java
@@ -23,7 +23,7 @@ public class BoardController {
     }
 
     @GetMapping
-    public Board getBoardById(Long id) throws Exception {
+    public Board getBoardById(Short id) throws Exception {
         Board board = boardService.getBoardById(id);
         if (board == null) {
             throw new Exception("Invalid board id");
