+++ src/main/java/com/ssafeople/backend/domain/board/service/BoardService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface BoardService {
     List<Board> getAllBoards();
 
-    Board getBoardById(Long id) throws Exception;
+    Board getBoardById(Short id) throws Exception;
 
     void updateBoardInfo(String boardName, String newDescription);
 }
